accordion functionality and styling: need to apply ng-deep tag to style shadow dom divs
