+++ apps/client/src/app/spec-download/spec-download.component.html
@@ -1,0 +1,11 @@
+<ngb-accordion #acc="ngbAccordion">
+  <ngb-panel title="Select an eRSD version to download">
+    <ng-template ngbPanelContent>
+      <span><button type="button" class="btn download-button" (click)="downloadJsonBundle()">Download Latest eRSD JSON Payload</button></span>
+        &nbsp;&nbsp;&nbsp;&nbsp;
+        <span><button type="button" class="btn download-button" (click)="downloadXmlBundle()">Download Latest eRSD XML Payload</button></span>
+        &nbsp;&nbsp;&nbsp;&nbsp;
+        <span><button type="button" class="btn download-button" (click)="downloadExcel()">Latest RCTC Spreadsheet & Change Log</button></span>
+    </ng-template>
+  </ngb-panel>
+</ngb-accordion>
